Moves openid service config to the new google subfolder
